--- src/mdx-pages/guild/membership.mdx
@@ -1,31 +1,0 @@
----
-title: "Jäsenyys"
-layout: "page"
----
-
-Voit hakea killan varsinaiseksi jäseneksi, mikäli olet AYY:n jäsen. Muussa tapauksessa voit hakea kiltaan vanhaksi jäseneksi tai ulkojäseneksi. Killan jäsenenä saat oikeuden osallistua killan tapahtumiin ja käyttää kiltahuonetta sekä pääsyn killan sähköpostilistoille ja Telegram-kanaville. Lisäksi varsinaisena jäsenenä saat äänioikeuden killan kokouksissa.
-
-# Uusi jäsen
-
-Uudeksi varsinaiseksi jäseneksi voit liittyä lukuvuosi-ilmoittautumisen yhteydessä tai täyttämällä alla olevan jäsenhakemuslomakkeen ja maksamalla jäsenmaksun killan tilille. 
-
-Killan vanhaksi jäseneksi voit hakea, jos olet ollut killan varsinainen jäsen, mutta et ole enää AYY:n jäsen. Vanhaksi jäseneksi voit hakea kirjeellä tai sähköpostilla killan hallitukselle. Vanhan jäsenen ei tarvitse maksaa jäsenmaksua.
-
-Killan ulkojäseneksi voidaan hyväksyä killan toiminnasta kiinnostunut henkilö, jota ei voida hyväksyä varsinaiseksi tai vanhaksi jäseneksi. Jäseneksi haetaan samoin kuin varsinaiseksi jäseneksi ja jäsenmaksun suuruus on sama.
-
-# Jäsenyyden uusiminen
-
-Jos olet jo ennestään killan jäsen ja haluat uusia jäsenyytesi, riittää että maksat jäsenmaksun lukuvuosi-ilmottautumisen yhteydessä tai killan tilille alla olevien ohjeiden mukaisesti. Jäsenmaksun eräpäivä on syyskuun viimeinen päivä, ja se oikeuttaa jäsenyyteen kyseiselle lukuvuodelle.
-
-# Jäsenmaksu
-
-Killan jäsenmaksu on 8€, ja sen voi maksaa joko lukuvuosi-ilmoittautumisen yhteydessä tai killan tilille seuraavilla tiedoilla:
-
-**Saaja:** Automaatio- ja systeemitekniikan kilta  
-**IBAN:** FI84 3131 3001 8081 61  
-**Viesti:** Etunimi + Sukunimi  
-**Summa:** 8 euroa  
-
-# Jäsenhakemus
-
-<iframe src="https://docs.google.com/forms/d/e/1FAIpQLSfTRrDzSNKXDLqmvuQp7EJswBgpeBMrvCPGqnI6P2Jl6IjQHA/viewform?embedded=true" width="100%" height="1500" frameBorder="0" marginHeight="0" marginWidth="0">Ladataan...</iframe>
